requirements.txt
-pencv-python==4.4.0.42
amqp==5.0.1
appdirs==1.4.4
arrow==0.16.0
asgiref==3.6.0
astroid==2.4.2
attrs==20.3.0
autocrop==1.0.2
Automat==20.2.0
autopep8==1.5.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
bibtexparser==1.2.0
billiard==3.6.3.0
cachetools==5.3.0
celery==5.0.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click-didyoumean==0.0.3
click-repl==0.1.6
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
defusedxml==0.7.1
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-bootstrap5==23.1
django-compat==1.0.15
django-crum==0.7.7
django-debug-toolbar==4.0.0
django-filter==2.3.0
django-ipware==3.0.1
django-model-utils==4.0.0
django-upload-validator==1.1.6
django-waffle==2.0.0
django-webpack-loader==0.7.0
Django==4.2
djangorestframework==3.11.1
drf-jwt==1.17.2
edx-django-utils==3.10.0
edx-drf-extensions==6.2.0
edx-opaque-keys==2.1.1
edx-proctoring==2.4.8
edx-rest-api-client==5.2.1
edx-when==1.3.0
event-tracking==1.0.0
fake-useragent==0.1.11
Flask==1.1.2
free-proxy==1.0.1
fs==2.4.11
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.17.3
googleapis-common-protos==1.59.0
httplib2==0.22.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.4.2
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonfield2==4.0.0.post0
kombu==5.0.2
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
newrelic==5.22.1.152
numpy==1.19.1
oauthlib==3.2.2
opencv-python==3.4.11.41
parsel==1.6.0
pbr==5.5.1
Pillow==7.2.0
pip==23.1.1
prettytable==0.7.2
progress==1.5
prompt-toolkit==3.0.8
Protego==0.1.16
protobuf==4.22.3
psutil==5.7.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.8
PyDispatcher==2.0.5
pygame==2.0.0.dev10
pyjwkest==1.4.2
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
requests-oauthlib==1.3.1
requests==2.24.0
rest-condition==1.0.3
rsa==4.9
rules==2.2
scholarly==0.4.1
scipy==1.4.1
Scrapy==2.4.1
selenium==3.141.0
semantic-version==2.8.5
service-identity==18.1.0
setuptools==56.0.0
six==1.15.0
slumber==0.7.1
soupsieve==2.0.1
spinners==0.0.24
sqlparse==0.3.1
stem==1.8.0
stevedore==3.2.2
tabulate==0.8.7
toml==0.10.1
tqdm==2.2.3
twisted-iocpsupport==1.0.1
Twisted==21.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.25.10
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
web-fragments==0.3.2
WebOb==1.8.6
Werkzeug==1.0.1
wrapt==1.12.1
XBlock==1.4.0
zope.interface==5.2.0
